+++ src/interfaces/drop_source_action.cpp
@@ -73,7 +73,7 @@ JsDropSourceAction::CreateNative(JSContext* cx)
 	return std::unique_ptr<JsDropSourceAction>(new JsDropSourceAction(cx));
 }
 
-size_t JsDropSourceAction::GetInternalSize()
+uint32_t JsDropSourceAction::GetInternalSize()
 {
 	return 0;
 }
