+++ src/interfaces/fb_metadb_handle.h
@@ -19,7 +19,7 @@ namespace mozjs
 		~JsFbMetadbHandle() override = default;
 
 		static std::unique_ptr<JsFbMetadbHandle> CreateNative(JSContext* cx, const metadb_handle_ptr& handle);
-		static size_t GetInternalSize(const metadb_handle_ptr& handle);
+		static uint32_t GetInternalSize(const metadb_handle_ptr& handle);
 
 	public:
 		metadb_handle_ptr& GetHandle();
