+++ src/interfaces/fb_profiler.h
@@ -21,7 +21,7 @@ namespace mozjs
 		~JsFbProfiler() override = default;
 
 		static std::unique_ptr<JsFbProfiler> CreateNative(JSContext* cx, const std::string& name);
-		static size_t GetInternalSize(const std::string& name);
+		static uint32_t GetInternalSize(const std::string& name);
 
 	public: // ctor
 		static JSObject* Constructor(JSContext* cx, const std::string& name = "");
