+++ src/interfaces/fb_ui_selection_holder.h
@@ -21,7 +21,7 @@ namespace mozjs
 		~JsFbUiSelectionHolder() override = default;
 
 		static std::unique_ptr<JsFbUiSelectionHolder> CreateNative(JSContext* cx, const ui_selection_holder::ptr& holder);
-		static size_t GetInternalSize(const ui_selection_holder::ptr& holder);
+		static uint32_t GetInternalSize(const ui_selection_holder::ptr& holder);
 
 	public:
 		void SetPlaylistSelectionTracking();
