+++ src/utils/text_helpers.h
@@ -5,10 +5,10 @@ namespace smp::utils
 	struct WrappedTextLine
 	{
 		std::wstring text;
-		size_t width;
+		uint32_t width;
 	};
 
-	[[nodiscard]] size_t GetTextHeight(HDC hdc, std::wstring_view text);
-	[[nodiscard]] size_t GetTextWidth(HDC hdc, std::wstring_view text, bool accurate = false);
-	[[nodiscard]] std::vector<WrappedTextLine> WrapText(HDC hdc, const std::wstring& text, size_t maxWidth);
+	[[nodiscard]] uint32_t GetTextHeight(HDC hdc, std::wstring_view text);
+	[[nodiscard]] uint32_t GetTextWidth(HDC hdc, std::wstring_view text, bool accurate = false);
+	[[nodiscard]] std::vector<WrappedTextLine> WrapText(HDC hdc, const std::wstring& text, uint32_t maxWidth);
 }
